refactor(components): reorganize component structure and update dashboard/status stats
