+++ src/components/componentsInventory/tables/ComponentTableComponent.vue
