+++ src/components/computers/tables/ComputerTableComponent.vue
