+++ src/components/reports/charts/ReportsAveragePriceChartComponent.vue
