+++ src/components/reports/charts/ReportsTypeCountChartComponent.vue
