+++ src/components/reports/sections/ReportsChartsComponent.vue
