+++ src/components/reports/sections/ReportsTableComponent.vue
