+++ src/components/ui/StatCardComponent.vue
