+++ src/views/dashboard/UsersView.vue
@@ -4,11 +4,11 @@
 // -------------------------------
 import type { CreateUserDTO } from '@dtos/user/CreateUserDTO';
 import type { EditUserDTO } from '@dtos/user/EditUserDTO';
-import ModalComponent from '@components/dashboard/ModalComponent.vue';
-import UserForm from '@components/dashboard/users/UserFormComponent.vue';
+import ModalComponent from '@components/ui/ModalComponent.vue';
+import UserForm from '@components/users/forms/UserFormComponent.vue';
 import type { UserInterface } from '@interfaces/UserInterface';
 import { UserService } from '@services/UserService';
-import UserTable from '@components/dashboard/users/UserTableComponent.vue';
+import UserTable from '@components/users/tables/UserTableComponent.vue';
 
 // -------------------------------
 // Third Party Imports
